+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'UltraCosmetics'
-include 'v1_20_R4', 'v1_21_R3', 'core', 'legacy-events', 'paper-support'
+include 'v1_20_R4', 'v1_21_R4', 'core', 'legacy-events', 'paper-support'
