--- v1_21_R3/build.gradle
@@ -1,6 +1,0 @@
-dependencies {
-    // https://hub.spigotmc.org/stash/projects/SPIGOT/repos/spigot/browse
-    implementation 'org.spigotmc:spigot:1.21.4-R0.1-SNAPSHOT:remapped-mojang'
-
-    implementation project(':core')
-}
