--- v1_21_R3/gradle.properties
@@ -1,2 +1,0 @@
-javaVersion=21
-remapMcVersion=1.21.4
