--- v1_21_R3/src/main/java/be/isach/ultracosmetics/v1_21_R3/ObfuscatedFields.java
@@ -1,12 +1,0 @@
-package be.isach.ultracosmetics.v1_21_R3;
-
-/**
- * This file automatically generated by obfuscatedfields_generator.py in tooling folder
- */
-public class ObfuscatedFields {
-    // corresponds to net.minecraft.world.entity.LivingEntity#jumping
-    public static final String JUMPING = "bm";
-
-    private ObfuscatedFields() {
-    }
-}
