--- v1_21_R3/src/main/java/be/isach/ultracosmetics/v1_21_R3/pets/CustomEntityPet.java
@@ -1,42 +1,0 @@
-package be.isach.ultracosmetics.v1_21_R3.pets;
-
-import be.isach.ultracosmetics.UltraCosmetics;
-import be.isach.ultracosmetics.cosmetics.pets.Pet;
-import be.isach.ultracosmetics.cosmetics.type.PetType;
-import be.isach.ultracosmetics.player.UltraPlayer;
-import be.isach.ultracosmetics.v1_21_R3.customentities.CustomEntities;
-import net.minecraft.world.entity.Entity;
-import org.bukkit.craftbukkit.v1_21_R3.entity.CraftEntity;
-import org.bukkit.entity.Mob;
-
-/**
- * @author RadBuilder
- */
-public abstract class CustomEntityPet extends Pet {
-
-    public CustomEntityPet(UltraPlayer owner, PetType petType, UltraCosmetics ultraCosmetics) {
-        super(owner, petType, ultraCosmetics);
-    }
-
-    @Override
-    public Mob spawnEntity() {
-        entity = (Mob) CustomEntities.spawnEntity(getNewEntity(), getPlayer().getLocation());
-        return entity;
-    }
-
-    @Override
-    protected void removeEntity() {
-        CustomEntities.removeCustomEntity(getNMSEntity());
-    }
-
-    @Override
-    public boolean isCustomEntity() {
-        return true;
-    }
-
-    public Entity getNMSEntity() {
-        return ((CraftEntity) entity).getHandle();
-    }
-
-    public abstract Entity getNewEntity();
-}
