+++ v1_21_R4/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    // https://hub.spigotmc.org/stash/projects/SPIGOT/repos/spigot/browse
+    implementation 'org.spigotmc:spigot:1.21.5-R0.1-SNAPSHOT:remapped-mojang'
+
+    implementation project(':core')
+}
