+++ v1_21_R4/src/main/java/be/isach/ultracosmetics/v1_21_R4/ObfuscatedFields.java
@@ -1,0 +1,12 @@
+package be.isach.ultracosmetics.v1_21_R4;
+
+/**
+ * This file automatically generated by obfuscatedfields_generator.py in tooling folder
+ */
+public class ObfuscatedFields {
+    // corresponds to net.minecraft.world.entity.LivingEntity#jumping
+    public static final String JUMPING = "bm";
+
+    private ObfuscatedFields() {
+    }
+}
