+++ v1_21_R4/src/main/java/be/isach/ultracosmetics/v1_21_R4/mount/MountSpider.java
@@ -1,0 +1,27 @@
+package be.isach.ultracosmetics.v1_21_R4.mount;
+
+import be.isach.ultracosmetics.UltraCosmetics;
+import be.isach.ultracosmetics.cosmetics.type.MountType;
+import be.isach.ultracosmetics.player.UltraPlayer;
+import be.isach.ultracosmetics.v1_21_R4.customentities.RideableSpider;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import org.bukkit.craftbukkit.v1_21_R4.CraftWorld;
+
+/**
+ * @author RadBuilder
+ */
+public class MountSpider extends MountCustomEntity {
+    public MountSpider(UltraPlayer owner, MountType type, UltraCosmetics ultraCosmetics) {
+        super(owner, type, ultraCosmetics);
+    }
+
+    @Override
+    public void onUpdate() {
+    }
+
+    @Override
+    public LivingEntity getNewEntity() {
+        return new RideableSpider(EntityType.SPIDER, ((CraftWorld) getPlayer().getWorld()).getHandle());
+    }
+}
